% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated-utilities.R
\name{call_api}
\alias{call_api}
\title{Call the World Bank API and return list}
\usage{
call_api(url_string, indicator)
}
\arguments{
\item{url_string}{A character string. A formatted url string}

\item{indicator}{A character string. indicatorID for request. Used for error returns}
}
\value{
json contents of page information
}
\description{
To be used inside of wbget()
}
