% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in end_points.R
\name{str.wblist}
\alias{str.wblist}
\title{Structure of a cache list}
\usage{
\method{str}{wblist}(object, max.level = 1, ...)
}

\arguments{
\item{object}{Cache list.}

\item{max.level}{`integer`. Passed to [utils::str()], defaults to `1` so
  only the top-level elements of the cache list are summarized.}

\item{...}{Further arguments passed to [utils::str()].}
}

\value{
`NULL`, invisibly. Called for the side effect of printing.
}

\description{
Structure of a cache list
}

\keyword{internal}
