% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wb_cachelist}
\alias{wb_cachelist}
\title{Cached information from the World Bank API}
\format{
An object of class \code{list} of length 8.
}
\usage{
data(wb_cachelist)
}
\description{
This data is a cached result of the \code{\link{wb_cache}} function.
By default functions \code{\link{wb_data}} and \code{\link{wb_search}} use this
data for the \code{cache} parameter.
}
\keyword{datasets}
