% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wb_data_flow.R
\name{wb_data_flow}
\alias{wb_data_flow}
\title{Download Data from the World Bank API (supports flows by counterpart area)}
\usage{
wb_data_flow(
  indicator,
  country = "countries_only",
  counterpart,
  sources,
  start_date,
  end_date,
  return_wide = TRUE,
  mrv,
  mrnev,
  cache,
  freq,
  gapfill = FALSE,
  date_as_class_date = FALSE,
  lang
)
}
\arguments{
\item{indicator}{Character vector of indicator codes. These codes correspond
to the \code{indicator_id} column from the \code{indicators} tibble of \code{\link[=wb_cache]{wb_cache()}}, \link{wb_cachelist}, or
the result of running \code{\link[=wb_indicators]{wb_indicators()}} directly}

\item{country}{Character vector of country, region, or special value codes for the
locations you want to return data for. Permissible values can be found in the
countries tibble in \link{wb_cachelist} or by running \code{\link[=wb_countries]{wb_countries()}} directly.
Specifically, values listed in the following fields \code{iso3c}, \code{iso2c}, \code{country},
\code{region}, \code{admin_region}, \code{income_level} and all of the \verb{region_*},
\verb{admin_region_*}, \verb{income_level_*}, columns. As well as the following special values
\itemize{
\item \code{"countries_only"} (Default)
\item \code{"regions_only"}
\item \code{"admin_regions_only"}
\item \code{"income_levels_only"}
\item \code{"aggregates_only"}
\item \code{"all"}
}}

\item{counterpart}{Character vector of a counterpart area id or name. Permissible values can be
found in the counterparts tibble in \link{wb_cachelist} or by running \code{\link[=wb_counterparts]{wb_counterparts()}} directly.
Note that there are several aggregate options:
*\code{"WLD"} (default - World)
*\code{"BND"} (Bondholders)
You can also filter for counterpart areas that are:
\itemize{
\item \code{"mdbs"}
\item \code{"bilateral"}
}}

\item{start_date}{Numeric or character. If numeric it must be in \verb{\%Y} form (i.e. four digit year).
For data at the subannual granularity the API supports a format as follows: for monthly data, "2016M01"
and for quarterly data, "2016Q1". This also accepts a special value of "YTD", useful for more frequently
updated subannual indicators.}

\item{end_date}{Numeric or character. If numeric it must be in \verb{\%Y} form (i.e. four digit year).
For data at the subannual granularity the API supports a format as follows: for monthly data, "2016M01"
and for quarterly data, "2016Q1".}

\item{return_wide}{Logical. If \code{TRUE} data is returned in a wide format instead of long,
with a column named for each \code{indicator_id} or if the \code{indicator} argument is a named vector,
the \code{\link[=names]{names()}} given to the indicator will be the column names. To necessitate this transformation,
the \code{indicator} column that provides the human readable description is dropped, but provided as a column label.
Default is \code{TRUE}}

\item{mrv}{Numeric. The number of Most Recent Values to return. A replacement
of \code{start_date} and \code{end_date}, this number represents the number of observations
you which to return starting from the most recent date of collection. This may include missing values.
Useful in conjuction with \code{freq}}

\item{mrnev}{Numeric. The number of Most Recent Non Empty Values to return. A replacement
of \code{start_date} and \code{end_date}, similar in behavior as \code{mrv} but excludes locations with missing values.
Useful in conjuction with \code{freq}}

\item{cache}{List of tibbles returned from \code{\link[=wb_cache]{wb_cache()}}. If omitted, \link{wb_cachelist} is used}

\item{freq}{Character String. For fetching quarterly ("Q"), monthly("M") or yearly ("Y") values.
Useful for querying high frequency data.}

\item{gapfill}{Logical. If \code{TRUE} fills in missing values by carrying forward the last
available value until the next available period (max number of periods back tracked will be limited by \code{mrv} number).
Default is \code{FALSE}}

\item{date_as_class_date}{Logical. If \code{TRUE} the date field is returned as class \link{Date}, useful when working with
non-annual data or data at mixed resolutions. Default is \code{FALSE}
available value until the next available period (max number of periods back tracked will be limited by \code{mrv} number).
Default is \code{FALSE}}

\item{lang}{Language in which to return the results. If \code{lang} is unspecified,
english is the default. For supported languages see \code{\link[=wb_languages]{wb_languages()}}.
Possible values of \code{lang} are in the \code{iso2} column. A note of warning, not
all data returns have support for langauges other than english. If the specific
return does not support your requested language by default it will return \code{NA}.}
}
\value{
a \link[tibble:tibble-package]{tibble} of all available requested data.
}
\description{
This function downloads the requested information using the World Bank API
}
\details{
\subsection{\code{obs_status} column}{

Indicates the observation status for location, indicator and date combination.
For example \code{"F"} in the response indicates that the observation status for
that data point is "forecast".
}
}
\examples{

# GDP of China (could also specify country as "CHN" the iso3c code)
\donttest{ids <- wb_data_flow("NY.GDP.MKTP.CD", country=c("China"))}

# GDP of Low and Middle Income countries
\donttest{ids <- wb_data_flow("NY.GDP.MKTP.CD", country=c("LMY"))}

# Multilateral amortization on debt owed to the World Bank IBRD by China
\donttest{ids <- wb_data_flow("DT.AMT.MLAT.CD", country=c("CHN"), counterpart=c("901"))}

#' # Multilateral amortization on debt owed to each of the Multilateral Development Banks by China
\donttest{ids <- wb_data_flow("DT.AMT.MLAT.CD", country=c("CHN"), counterpart=c("mdbs"))}

# Multilateral and bilateral amortization on debt owed to any creditor by China
\donttest{ids <- wb_data_flow(c("DT.AMT.MLAT.CD", "DT.AMT.BLAT.CD"), country=c("CHN"), counterpart=c("WLD"))}

}
