% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_cache.R
\name{wb_default_lang}
\alias{wb_default_lang}
\title{Check the Default Langauge for World Bank API calls}
\usage{
wb_default_lang(lang)
}
\arguments{
\item{lang}{Language in which to return the results. If \code{lang} is unspecified,
english is the default. For supported languages see \code{\link[=wb_languages]{wb_languages()}}.
Possible values of \code{lang} are in the \code{iso2} column. A note of warning, not
all data returns have support for langauges other than english. If the specific
return does not support your requested language by default it will return \code{NA}.}
}
\value{
character
}
\description{
Check the Default Langauge for World Bank API calls
}
