% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in wb_variable_coverage.R
\name{wb_indicator_coverage}
\alias{wb_indicator_coverage}
\title{Summarize Data Coverage for World Bank Indicators}
\usage{
wb_indicator_coverage(indicator)
}

\arguments{
\item{indicator}{A character string (e.g., \code{"SP.DYN.LE00.IN"}).}
}

\value{
A \code{data.table} with one row per country.
}

\description{
Given a indicator available from the World Bank API, this functions returns a pre-computed summary stored on GitHub and not available from the API that shows the percentage of non-missing (non-\code{NA}) observations for every country, the year range per country, and the total number of observations per country. This is useful before downloading a dataset.
}

\examples{
\dontrun{
my_indicators <- c(
  life_exp   = "SP.DYN.LE00.IN",
  gdp_capita = "NY.GDP.PCAP.CD",
  pop        = "SP.POP.TOTL"
)

# summary for each indicator
wb_indicator_coverage(my_indicators)
}
}

