% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in wb_variable_coverage.R
\name{wb_variable_coverage}
\alias{wb_variable_coverage}
\title{Summarize Data Coverage for World Bank Indicators}
\usage{
wb_variable_coverage(variable)
}

\arguments{
\item{variable}{A character string (e.g., \code{"SP.DYN.LE00.IN"}).}
}

\value{
A \code{data.table} with one row per country.
}

\description{
Given a variable available from the World Bank API, this functions returns a pre-computed summary stored on GitHub and not available from the API that shows the percentage of non-missing (non-\code{NA}) observations available per variable for every country, the year range per country, and the total number of observations per country. This is useful before downloading the actual indicators.
}

\examples{
\dontrun{

# summary for each variable
wb_variable_coverage(d)
}
}

