% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated-wbcache.R
\name{wbcountries}
\alias{wbcountries}
\title{Download updated country and region information from World Bank API}
\usage{
wbcountries(lang = c("en", "es", "fr", "ar", "zh"))
}
\arguments{
\item{lang}{Language in which to return the results. If \code{lang} is unspecified,
english is the default.}
}
\value{
A data frame of available countries and regions with related information
}
\description{
Download updated information on available countries and regions
from the World Bank API
}
\note{
Not all data returns have support for langauges other than english. If the specific return
does not support your requested language by default it will return \code{NA}. For an enumeration of
supported languages by data source please see \code{\link{wbdatacatalog}}.
The options for \code{lang} are:
\itemize{
\item \code{en}: English
\item \code{es}: Spanish
\item \code{fr}: French
\item \code{ar}: Arabic
\item \code{zh}: Mandarin
}
}
