% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated-wbcache.R
\name{wbdatacatalog}
\alias{wbdatacatalog}
\title{Download an updated list of the World Bank data catalog}
\usage{
wbdatacatalog()
}
\value{
A data frame of the World Bank data catalog with related information
}
\description{
Download an updated list of the World Bank data catalog
from the World Bank API
}
\note{
This function does not support any languages other than english due to
the lack of support from the World Bank API
}
