% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utilities.R
\name{wbformatcols}
\alias{wbformatcols}
\title{Format column names of World Bank API returns}
\usage{
wbformatcols(df, col_names, blank2NA = TRUE)
}
\arguments{
\item{df}{data frame returned from API call}

\item{col_names}{a named vector of column names and what they will
be changed to}

\item{blank2NA}{logical. If \code{TRUE} blank values are converted to NA}
}
\value{
a data frame with the column names changed accordingly
}
\description{
change the name of columns that we know of to our
standardized names. If there are new columns added that we
don't know about, leave their name the same
}

