% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{wbget.raw}
\alias{wbget.raw}
\title{Call the World Bank API and return list}
\usage{
wbget.raw(url)
}
\arguments{
\item{url}{A charcter string. A formatted url string}
}
\value{
A list with page information and the data itself
}
\description{
To be used inside of wbget()
}
