% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{wbget}
\alias{wbget}
\title{Call the World Bank API and return a formatted data frame}
\usage{
wbget(url)
}
\arguments{
\item{url}{A charcter string. A formatted url string without page information}
}
\value{
A data frame
}
\description{
This function calls wbget.raw and determines if there are
multiple pages from the request. If there are multiple pages
then if generates a list of data frames from each page and
then combines the results with do.call("rbind", mylist)
}
