% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated-utilities.R
\name{wbget_dc}
\alias{wbget_dc}
\title{Call the Data Catalog API}
\usage{
wbget_dc(url_string)
}
\arguments{
\item{url_string}{A charcter string. A formatted url string}
}
\value{
A list of data frames
}
\description{
Helper function for the data catalog call
}
\note{
This call is seperate because the data catalog is actaully a different
 API and therefore has a different return structure.
}
