% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in wbstats-package.R
\name{wbstats}
\alias{wbstats}
\title{Programmatic Access to Data and Statistics from the World Bank API}
\description{
Search and download data from the World Bank Data API. Includes
 support for mutliple languages, access to annual, quarterly, and monthly
 data.
}

