% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{wburls}
\alias{wburls}
\title{url chucks to be used in API calls}
\usage{
wburls()
}
\value{
a list with a base url and a url section for formatting the json return
}
\description{
This function is used inside other functions in this package
}
