% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in data.R
\name{wb_cachelist}
\alias{wb_cachelist}
\title{Cached information from the World Bank API}
\description{
This data is a cached result of the \code{\link{wb_cache}} function.
By default functions \code{\link{wb_data}} and \code{\link{wb_search}} use this
data for the \code{cache} parameter.
}

