% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in wb_country_coverage.R
\name{wb_country_coverage}
\alias{wb_country_coverage}
\title{Summarize Data Coverage for World Bank Indicators}
\usage{
wb_country_coverage(pattern, country, start_date, end_date, cache)
}

\arguments{
\item{pattern}{A character string or regular expression (e.g., \code{"GDP"}).}

\item{country}{Character vector of country, region, or special value codes for the locations you want to return data for. Permissible values can be found in the countries data frame in \link{wb_cachelist} or by running \code{\link[=wb_countries]{wb_countries()}} directly. Specifically, values listed in the following fields \code{iso3c}, \code{iso2c}, \code{country}, \code{region}, \code{admin_region}, \code{income_level} and all of the \code{region_*}, \code{admin_region_*}, \code{income_level_*}, columns. As well as the following special values

\itemize{
\item \code{"countries_only"} (Default)
\item \code{"regions_only"}
\item \code{"admin_regions_only"}
\item \code{"income_levels_only"}
\item \code{"aggregates_only"}
\item \code{"all"}
}}

\item{start_date}{Numeric or character. If numeric it must be in \code{\%Y} form (i.e. four digit year). For data at the subannual granularity the API supports a format as follows: for monthly data, "2016M01" and for quarterly data, "2016Q1". This also accepts a special value of "YTD", useful for more frequently updated subannual indicators.}

\item{end_date}{Numeric or character. If numeric it must be in \code{\%Y} form (i.e. four digit year). For data at the subannual granularity the API supports a format as follows: for monthly data, "2016M01" and for quarterly data, "2016Q1".}

\item{cache}{List of data frames returned from \code{\link[=wb_cache]{wb_cache()}}. If omitted, \link{wb_cachelist} is used}
}

\value{
A \code{data.table} with one row per country.
}

\description{
Given a indicator available from the World Bank API, this functions returns a pre-computed summary stored on GitHub and not available from the API that shows the percentage of non-missing (non-\code{NA}) observations for every country, the year range per country, and the total number of observations per country. This is useful before downloading a dataset.
}

\examples{
\dontrun{
wb_country_coverage("gross domestic product", c("Mexico", "Chile"), 2010, 2020)
}
}

