% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wb_data.R
\name{wb_data}
\alias{wb_data}
\title{Download Indicator Data from World Bank}
\usage{
wb_data(indicator = "SP.POP.TOTL", country = "AFG", start_date,
  end_date, mrv, freq, mrnev, gapfill, footnote, scale, cache)
}
\arguments{
\item{scale}{}
}
\description{
Download Indicator Data from World Bank
}
