% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{wb_search}
\alias{wb_search}
\title{Search indicator information available through the World Bank API}
\usage{
wb_search(
  pattern = "poverty",
  fields = c("indicator", "indicator_desc"),
  extra = FALSE,
  cache,
  ...
)
}
\arguments{
\item{pattern}{Character string or regular expression to be matched}

\item{fields}{Character vector of column names through which to search}

\item{extra}{if \code{FALSE}, only the indicator ID and short name are returned,
if \code{TRUE}, all columns of the \code{cache} parameter's indicator data frame
are returned}

\item{cache}{List of data frames returned from \code{\link{wb_cache}}. If omitted,
\code{\link{wb_cachelist}} is used}
}
\value{
Data frame with indicators that match the search pattern.
}
\description{
This function allows finds indicators that match a search term and returns
a data frame of matching results
}
\examples{
wb_search(pattern = "education")

wb_search(pattern = "Food and Agriculture Organization", fields = "source_org")

# with regular expression operators
# 'poverty' OR 'unemployment' OR 'employment'
wb_search(pattern = "poverty|unemployment|employment")

 # pass any other grep argument along as well
 # everything without 'education'
 wb_search(pattern = "education", invert = TRUE)
}
