% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated-utilities.R
\name{wbget}
\alias{wbget}
\title{Call the World Bank API and return a formatted data frame}
\usage{
wbget(url_string, indicator)
}
\arguments{
\item{url_string}{A character string. A formatted url string}

\item{indicator}{A character string. indicatorID for request. Used for error returns}
}
\value{
A data frame
}
\description{
This function calls wbget.raw and determines if there are
multiple pages from the request. If there are multiple pages
then if generates a list of data frames from each page and
then combines the results with do.call("rbind", mylist)
}
