% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/wbstats.R
\docType{package}
\name{wbstats}
\alias{wbstats}
\alias{wbstats-package}
\title{wbstats: A package for programmatically accessing the World Bank API.}
\description{
The wbstats package provides structured access to all of the information
available from the World Bank API including; support for mutliple languages,
access to all annual, quarterly, and monthly data, as well as access to the
World Bank Data Catalog API, which provides metadata on each data source,
such as update schedules, last revisions, and coverage, among other information.
}
\details{
To learn more about the wbstats package, start with the vignettes: \code{browseVignettes(package = "wbstats")}
}

