% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wbstats-package.R
\docType{package}
\name{wbstats}
\alias{wbstats}
\title{wbstats: An R package for searching and downloading data from the World Bank API.}
\description{
The wbstats package provides structured access to data
available from the World Bank API including; support for mutliple languages,
access to annual, quarterly, and monthly data.
}
